graphics/lab1/requirements.txt
glfw==2.7.0
numpy==1.26.4
pygame==2.5.2
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
